+++ src/Blueprint/Blueprint/ChangeLog.md
@@ -18,6 +18,13 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 0.2.2
+- Add change note support during publishing of a blueprint
+- Update Blueprint SDK version
+- Bug fixes
+
+## Version 0.2.1
 * Bug fixes and improvements
 
 ## Version 0.2.0